+++ episimlab/utils/__init__.py
@@ -1,4 +1,4 @@
-from .discrete_time_approx import *
-from .midx import *
 from .datetime import *
 from .variable import *
+from .profile import *
+from .plot import *
